Fix some warnings

+++ tvipt/util.h
@@ -42,13 +42,13 @@ inline bool stream_copy_breakable(Stream &src, Stream &dst, uint16_t max_bytes, 
 }
 
 // Find the index of the value of the property of an object whose name matches the specified string
-inline int find_json_prop(const char *json, jsmntok_t *tokens, size_t num_tokens, int object, const char *prop_name) {
+inline int find_json_prop(const char *json, jsmntok_t *tokens, int num_tokens, int object, const char *prop_name) {
     // Easy way: scan through all tokens looking for parentage
     for (int i = 0; i < num_tokens; i++) {
         jsmntok_t *tok = &tokens[i];
         if (tok->parent == object) {
             if (strlen(prop_name) == tok->end - tok->start &&
-                strncmp(json + tok->start, prop_name, tok->end - tok->start) == 0) {
+                strncmp(json + tok->start, prop_name, (size_t) tok->end - tok->start) == 0) {
                 // The next token is the value
                 return i + 1;
             }
